# Conflicts:
#	Podfile.lock
#	Pods/Manifest.lock
#	Pods/Pods.xcodeproj/project.pbxproj
#	Pods/Target Support Files/Pods-Q-municate/Pods-Q-municate-resources.sh
#	Pods/Target Support Files/Pods-Q-municate/Pods-Q-municate.debug.xcconfig
#	Pods/Target Support Files/Pods-Q-municate/Pods-Q-municate.release.xcconfig
#	Q-municate.xcodeproj/project.pbxproj

+++ Q-municate/Scenes/Settings/QMUpdateUserViewController/QMUpdateUserViewController.m
@@ -215,12 +215,12 @@ replacementString:(NSString *)string  {
     
     if (alreadyExpanded) {
         [self reloadExpandedCellForIndexPath:cellIndexPath
-                                    duration:duration
+                                    duration:(CGFloat)duration
                             withRowAnimation:UITableViewRowAnimationNone];
     }
     else {
         [self expandCellForIndexPath:cellIndexPath
-                            duration:duration
+                            duration:(CGFloat)duration
                     withRowAnimation:UITableViewRowAnimationAutomatic];
     }
 }
